GH-3433 remove obselete base classes/interfaces
fixup! Merge pull request #3381 from JervenBolleman/GH-3350-action-test-with-java-17-develop
