--- core/common/iterator/src/main/java/org/eclipse/rdf4j/common/iteration/CloseableIterationBase.java
@@ -1,17 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Distribution License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *******************************************************************************/
-package org.eclipse.rdf4j.common.iteration;
-
-/**
- * @deprecated since 2.0. Use {@link AbstractCloseableIteration} instead.
- * @author Jeen Broekstra
- */
-@Deprecated
-public abstract class CloseableIterationBase<E, X extends Exception> extends AbstractCloseableIteration<E, X> {
-
-}
