+++ core/http/client/src/main/java/org/eclipse/rdf4j/http/client/HttpClientSessionManager.java
@@ -15,30 +15,26 @@ import org.apache.http.client.HttpClient;
  * @author James Leigh
  */
 @SuppressWarnings("deprecation")
-public interface HttpClientSessionManager extends SesameClient {
+public interface HttpClientSessionManager {
 
 	/**
 	 * @return Returns the httpClient.
 	 */
-	@Override
 	HttpClient getHttpClient();
 
 	/**
 	 * Creates a new SPARQL Protocol session to the remote SPARQL endpoint.
 	 */
-	@Override
 	SPARQLProtocolSession createSPARQLProtocolSession(String queryEndpointUrl, String updateEndpointUrl);
 
 	/**
 	 * Creates a new session to the remote RDF4J REST API.
 	 */
-	@Override
 	RDF4JProtocolSession createRDF4JProtocolSession(String serverURL);
 
 	/**
 	 * Closes any remaining connections and threads used by the sessions created by this object.
 	 */
-	@Override
 	void shutDown();
 
 }
