--- core/sail/api/src/main/java/org/eclipse/rdf4j/sail/helpers/SailBase.java
@@ -1,17 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Distribution License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *******************************************************************************/
-package org.eclipse.rdf4j.sail.helpers;
-
-/**
- * @deprecated since 4.0. Use {@link AbstractSail} instead.
- * @author Jeen Broekstra
- */
-@Deprecated
-public abstract class SailBase extends AbstractSail {
-
-}
