+++ core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/Stats.java
@@ -12,8 +12,8 @@ import org.eclipse.rdf4j.common.annotation.InternalUseOnly;
 import org.eclipse.rdf4j.model.Statement;
 
 /**
- * @deprecated since 3.0. This feature is for internal use only: its existence, signature or behavior may change without
- *             warning from one release to the next.
+ * @apiNote This feature is for internal use only: its existence, signature or behavior may change without warning from
+ *          one release to the next.
  */
 @InternalUseOnly
 public class Stats {
