+++ core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/results/lazy/ValidationResultIterator.java
@@ -17,7 +17,7 @@ import java.util.Set;
 
 import org.eclipse.rdf4j.common.annotation.InternalUseOnly;
 import org.eclipse.rdf4j.common.iteration.CloseableIteration;
-import org.eclipse.rdf4j.common.iteration.IteratorCloseableIteration;
+import org.eclipse.rdf4j.common.iteration.CloseableIteratorIteration;
 import org.eclipse.rdf4j.sail.SailException;
 import org.eclipse.rdf4j.sail.shacl.ast.planNodes.ValidationTuple;
 import org.eclipse.rdf4j.sail.shacl.results.ValidationResult;
@@ -104,8 +104,7 @@ public class ValidationResultIterator implements Iterator<ValidationResult> {
 			actualList.add(tupleIterator.next());
 		}
 
-		tupleIterator = new IteratorCloseableIteration<>(actualList.iterator());
-
+		tupleIterator = new CloseableIteratorIteration<>(actualList.iterator());
 		return Collections.unmodifiableList(actualList);
 	}
