+++ core/spin/src/main/java/org/eclipse/rdf4j/spin/SpinRenderer.java
@@ -104,7 +104,7 @@ import org.eclipse.rdf4j.query.parser.sparql.SPARQLUpdateDataBlockParser;
 import org.eclipse.rdf4j.rio.RDFHandler;
 import org.eclipse.rdf4j.rio.RDFHandlerException;
 import org.eclipse.rdf4j.rio.RDFParseException;
-import org.eclipse.rdf4j.rio.helpers.RDFHandlerBase;
+import org.eclipse.rdf4j.rio.helpers.AbstractRDFHandler;
 
 import com.google.common.base.Function;
 
@@ -1227,7 +1227,7 @@ public class SpinRenderer {
 		private void renderDataBlock(String data) throws RDFHandlerException {
 			SPARQLUpdateDataBlockParser parser = new SPARQLUpdateDataBlockParser(valueFactory);
 			parser.setAllowBlankNodes(false); // no blank nodes allowed
-			parser.setRDFHandler(new RDFHandlerBase() {
+			parser.setRDFHandler(new AbstractRDFHandler() {
 
 				final Map<Resource, ListContext> namedGraphLists = new HashMap<>();
